+++ src/lib/client/SuperDebug.svelte
@@ -1,3 +1,5 @@
+<svelte:options runes={false} />
+
 <script>
 	import { browser } from '$app/environment';
 	import { page } from '$app/stores';