+++ src/lib/adapters/typebox.ts
@@ -6,18 +6,41 @@ import {
 	type ValidationResult,
 	type ClientValidationAdapter
 } from './adapters.js';
-import type { TSchema } from '@sinclair/typebox';
-import type { TypeCheck } from '@sinclair/typebox/compiler';
+import type { TSchema } from 'typebox';
+import type { Validator } from 'typebox/compile';
 import { memoize } from '$lib/memoize.js';
+import Type from 'typebox';
+
+/**
+ * Custom TypeBox type for Date for testing purposes.
+ */
+export class TDate extends Type.Base<globalThis.Date> {
+	public override Check(value: unknown): value is globalThis.Date {
+		return value instanceof globalThis.Date;
+	}
+	public override Errors(value: unknown): object[] {
+		return this.Check(value) ? [] : [{ message: 'must be Date' }];
+	}
+	public override Create(): globalThis.Date {
+		return new globalThis.Date(0);
+	}
+}
+
+/**
+ * Custom TypeBox type for Date for testing purposes.
+ */
+export function Date(): TDate {
+	return new TDate();
+}
 
 // From https://github.com/sinclairzx81/typebox/tree/ca4d771b87ee1f8e953036c95a21da7150786d3e/example/formats
 const Email =
 	/^[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/i;
 
 async function modules() {
-	const { TypeCompiler } = await import(/* webpackIgnore: true */ '@sinclair/typebox/compiler');
-	const { FormatRegistry } = await import(/* webpackIgnore: true */ '@sinclair/typebox');
-	return { TypeCompiler, FormatRegistry };
+	const { Compile } = await import(/* webpackIgnore: true */ 'typebox/compile');
+	const Format = await import(/* webpackIgnore: true */ 'typebox/format');
+	return { Compile, Format };
 }
 
 const fetchModule = /* @__PURE__ */ memoize(modules);
@@ -26,14 +49,14 @@ async function validate<T extends TSchema>(
 	schema: T,
 	data: unknown
 ): Promise<ValidationResult<Infer<T, 'typebox'>>> {
-	const { TypeCompiler, FormatRegistry } = await fetchModule();
+	const { Compile, Format } = await fetchModule();
 
 	if (!compiled.has(schema)) {
-		compiled.set(schema, TypeCompiler.Compile<TSchema>(schema));
+		compiled.set(schema, Compile(schema));
 	}
 
-	if (!FormatRegistry.Has('email')) {
-		FormatRegistry.Set('email', (value) => Email.test(value));
+	if (!Format.Has('email')) {
+		Format.Set('email', (value: string) => Email.test(value));
 	}
 
 	const validator = compiled.get(schema);
@@ -46,7 +69,7 @@ async function validate<T extends TSchema>(
 	return {
 		success: false,
 		issues: errors.map((issue) => ({
-			path: issue.path.substring(1).split('/'),
+			path: issue.instancePath ? issue.instancePath.substring(1).split('/') : [],
 			message: issue.message
 		}))
 	};
@@ -74,4 +97,4 @@ function _typeboxClient<T extends TSchema>(
 export const typebox = /* @__PURE__ */ memoize(_typebox);
 export const typeboxClient = /* @__PURE__ */ memoize(_typeboxClient);
 
-const compiled = new WeakMap<TSchema, TypeCheck<TSchema>>();
+const compiled = new WeakMap<TSchema, Validator<{}, TSchema>>();