+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Possibly fixed the SuperDebug broken import on Svelte 5 in enforced runes mode (#599)[https://github.com/ciscoheat/sveltekit-superforms/issues/599]
+
 ## [2.27.4] - 2025-10-14
 
 ### Security